Fix terraform destroy
To really get rid of a domain the code should "undefine" it instead of
calling the "destroy" function. The "destroy" code is about forcefully
shut down a domain, it's not about removing it.

This seemed to work in the past because the domain were automatically
undefined by libvirt, see PR #26.

This code avoids to raise an error when the IP address of an interface
is not found, this is needed when trying to destroy a domain which is
shut down.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>