Discover IP address of the network cards attached to a LAN
When a network card is attached to a LAN the `domain.ListAllInterfaceAddresses`
API cannot be used to get the IP address associated with the interfaces.

The only solution to this problem is to use the [Qemu guest agent](http://wiki.libvirt.org/page/Qemu_guest_agent)
to fetch this information.

This PR will take advantage of the qemu guest ageint (when installed
inside of the domain) to workaround the limitations of the
`domain.ListAllInterfaceAddresses`.

In theory this PR could also allow to relax the required version of
libvirt.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>