Better handling of errors while creating a domain
Sometimes if an error occurs at domain creation time, the resulting
domain is **unknown** to terraform. That makes impossible to remove
it, causing later runs of `terraform apply` to fail.

This commit marks the domain ID as a partial attribute that has to be
saved to the `.tfstate` as soon as it's created.

That should fix issues like #67