CoreOS Ignition support updates
This patch updates the CoreOS ignition support. "coreos_ignition"
is now a block, that looks as follows:

   coreos_ignition {
     content = <string> *Required*
     generated_ignition_file_dir = <string> *Optional*
   }

If the "coreos_ignition" block is present, then content must be
defined and can point to either the location of an existing
ignition file or the contents of a Terraform Ignition provider
object.  If the latter, you can also opitionally specify the
directory in which the ignition file will be written using
"generated_ignition_file_dir".  If this directory isn't specified,
"/tmp" will be used.