Add network option to disable DHCP services
When creating a VM with multiple network interfaces on different
networks, you typically don't want a DHCP server running on every
network. For example, a CoreOS VM will send out a DHCP request on every
network interface by default when it boots up (given no other config
from say, Ignition). This is problematic for the VM's networking
because it will end up with multiple default routes.

An example use case of this is a VM with DHCP enabled on one NIC (to be
used as the "management" network), from which the rest of the VM's
interfaces will be configured.