Write Ignition file as a volume in a libvirt storage pool
This avoids the problem where the Ignition file is remote to
the host on which the libvirt domain is being created.  We've
added a "libvirt_ignition" resource which manages the Ignition
file in the libvirt volume - creates it, and destroys it.  The
"coreos_ignition" field in the libvirt_domain definition must
point to the Id of a "libvirt_ignition" resource.

The code is modelled on that used for CloudInit.