Add new method to get information on domain interfaces
We've found that, when executing destroy/create cycles on
a domain where the virtual network(s) to which it is attached
are persisted across these cycles then we get network failures
on recreation of the domain since the old entries for
the domain are still present in the virtual network(s). This issue
should (hopefully) be fixed by commit
a20d2a92668b2d5e077267c50074bf853ef371d.  We've also found that,
on occasion, we don't get complete information on a domain's
interfaces from ListAllInterfaceAddresses. This patch implements
a new method getDomainInterfacesFromNetworks which is called from
domainGetIfacesInfo after the qemu agent method has been used if
the qemu agent method doesn't return any information.  This new
method builds the interface information from information collected
from the networks themselves.

Change-Id: I5271ee191db93b2b1a0f14dc226d0b945f77a901