Add scsi controller option and wwn to virtual disks
We've added the facility to define a disk bus type of scsi to the
specification of a disk in the domain definition.  If the disk is
a scsi disk, a random wwn is generated unless a wwn is provided for
that disk.

The disk stanza now looks as follows:

   disk {
     volume_id = "${libvirt_volume.mydisk.id}"
     scsi = "yes"
     wwn = "05abcd123456789a"
   }

Having "scsi" present with any value will specify a scsi bus for
the disk.  If "wwn" is present for a scsci disk then the value of
"wwn" is used for the disk wwn, otherwise a random value is generated.