Dynamically start DHCP services on networks
Rather than starting DHCP services on a network when it is created, they
should be started dynamically when domains are created with interfaces
on the network with requested addresses (or hostnames).

Because domains can be created in parallel, some read/write access to
the libvirt network definition must be synced to prevent an error that
occurs when a DHCP range is added twice to a network.