Fix broken tests
The tests are failing because travis has an old version of libvirt that
is not happy about the xml we generate.

Unfortunately there's no way to get a more recent version of ubuntu on
travis (it's stuck with trusty!!) and there are not libvirt 2.0
packages for trusty.

This commit fixes the issues in the "inception" way:

  * travis provisions a ubuntu trusty VM as usual
  * the travis/host script does the following steps:
    * installs LXD
    * creates a system container based on the latest release of ubuntu
    * runs the container as privileged one
    * mount the git checkout into the container as read-only mount
      point (it easier than having a RW one)
    * runs the travis/guest script inside of the container:
      * install libvirt + go 1.8
      * run the acceptance tests, save coverage to file
  * on the travis VM:
    * pull the coverage results file from the container
    * send it to coverall

So far everything is working and is a miracle given the container,
based on a more recent version of ubuntu, is using the kernel of the
host, the old ubuntu trusty.

This is just a termporary workaround until either travis updates its
VMs to a more recent release of ubuntu or we switch the project to use
Jenkins with an openSUSE worker.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>