Adds fuller support for arch and machine settings Allows updates to be tracked Sets sensible defaults
Due to the way libvirt works if you specific machine types like pc,isapc
or q35 these get translated into 'canonical' forms, ie. the latest
version

When quering terraform will notice a change, as it would still be
looking for pc and will get something like pc-i440fx-2.9.
Code has been added to assume that if we see 'pc-i440fx-2.9' and it is
canonical for pc then we translate back to pc for terraform.
This could cause issues if you are specifically using pc-i440fx-2.9 as
the machine type in your terraform file.

The only other alternative was to translate pc > pc-i440fx-2.9 in
defDomain which would require a libvirt connection when one often would
be available, this felt very hacky.