FIX waitForLeases invalid key issue
We are modifying network interface object after adding it as key to
waitForLeases map. This makes the interface object stored in the key
invalid when it is used in functions like hasNetworkAddress() later.

As a result of this the code does not wait for the interface to get
the lease even after specifying `wait_for_lease = true` in the tf
file.

To fix this we move the code to add interface struct object to
waitForLeases map only after all modifications are done i.e. near
to the end of for loop.