Full support for graphics and video settings
This patch adds full support for vnc and spice graphics types
as well as support for changing the video device.

By default a graphics device will no longer be added, this is due
to older code assuming that the graphics type would never be read or
changed. As two of the architectures the code supports are ppc64 and
s390x and these don't support graphics a default wouldn't work.

VNC support now includes setting the listen port and listen address