Expose IP addresses for domains at the top level
Accessing a list, nested inside a map, nested inside a list, is very
difficult in Terraform, so the existing structure makes it impossible to
extract IP addresses for all domains, when combined with `count` in a
domain definition.

Exposing `ip_address` on the domain makes IP addresses much more
accessible.