Rework dns forwarders and add localonly dns option to libvirt-network
I make a new DNS a sectrion of network, more like the upstream xml. Inside
that new DNS section I include forwarders, which were under the main object,
so this is a breaking change. But one that aligns with the upstream xml and
one which allows much easier future growth of DNS management.

I add 'localOnly' which tells dnsmasq to own all of the domain in question.
Instead of forwarding unknown names. This way you can point your local
DNS at the libvirt DNS for the domain in question, and the domain in
question will not forward unknown hosts back to your local DNS.