Fix incompatible change from 77982e
https://github.com/dmacvicar/terraform-provider-libvirt/commit/77982e7e171afb8354915ef286fa025f5ee77506#diff-2f03be069459ebc91fcf0607d4d686d8R832
checks if the source network is a bridge but doesn't set the bridge
value correctly in the Terraform state just after.

See https://godoc.org/github.com/libvirt/libvirt-go-xml#DomainInterfaceSourceNetworkhttps://godoc.org/github.com/libvirt/libvirt-go-xml#DomainInterfaceSourceBridge
for reference.

Also, in that case, .Source.Network is not defined so better let it to the
default value.