domain: telnet console support
Add a new console connection type - telnet over tcp. This feature is
useful when LAN access to the domain is required but the guest OS is
not provisioned with networking.

Treat all other console types as they were pty based. This is needed
to avoid regression with unknown types since no value bounds check
was in place.

As the telnet default port is a system port, use cisco-sccp service
in the tests. This service name is IANA assigned, is not likely to
be seen on a Linux and resolves on a wide set of distros.