Use config-drive instead of guestfish on s390x (#678)
* Revert "Create guestfish upload directory for coreos ignition"

This reverts commit 60a702f5c98211f11ef453a8edd5e44a14bac4d6.

Instead of using libguestfish to inject the Ignition configs directly
into the root disk of the virtual machine, we are going to use the
OpenStack config-drive mechanism to pass the Ignition config through.
Since this code is no longer needed, I'm removing it.

* Revert "Add support for s390x on coreos ignition"

This reverts commit ce1aa007ce514ef2b045386ebd0d280e5f719263.

Instead of using libguestfish to inject the Ignition configs directly
into the root disk of the virtual machine, we are going to use the
OpenStack config-drive mechanism to pass the Ignition config through.
Since this code is no longer needed, I'm removing it.

* Add support for s390x to cloud-init config drive

In order for an OpenStack config-drive to be attached in s390x, it needs
to use a different set of parameters (e.g. SCSI vs IDE).

* Add support for s390x to Ignition configs

This allows Ignition configs to be passed to s390x virtual machines
using the OpenStack config-drive. In order for this to work, the
Ignition instance within the guest needs to be instructed to use the
OpenStack provider.

* Document s390x support

I don't imagine that many people will have a mainframe laying around,
but just in case, here are the instructions for getting Ignition to work
inside of an s390x guest.