Use virtio-blk instead of disk for injecting ignition config for s390x/ppc64le
The method of mimicing what Openstack does for injecting ignition config works for
images which have the provider as Openstack because ignition recognizes the platform
and knows it has to get the ignition config from the config drive. For QEMU images, ignition
supports getting the config from the firmware config device which is not supported by ppc64
and s390x.

The workaround we have used thus far is to use the Openstack image on the QEMU platform but have
this provider create the iso containing the ignition config. There was a discussion in ignition
(https://github.com/coreos/ignition/issues/928) to have a more QEMU based method of injecting ignition
config and it was decided to use a virtio-blk device with a serial of `ignition` which ignition can
recognize. This was mainly because with external devices, it is hard to tell if there is an issue with
the device or if the kernel has not detected it yet if it has a long discovery phase.

This PR reverts the ISO method used by Openstack and just creates a virtio-blk device through the QEMU
command line options.