Set firmware attribute to efi for aarch64
This new firmare attribute was added to newer versions of libvirt-go-xml. This enables to specify the firmware as
'bios' or 'efi' and automatically fills in the loader and NVRAM elements without having to specify them manually.
Refer to https://libvirt.org/formatdomain.html#bios-bootloader for this. As part of this change, libvirt-go and
libvirt-go-xml needs to be updated.

Setting this automatically for aarch64 as it is a EFI architecture by default.