Support dnsmasq options
So this is implementing #642 using datasource as suggested in the issue. The issue was closed at that time because
of a workaround using xslt. But now we are seeing that this is needed for our Openshift CI for non-x86 arches which
runs on libvirt. This would enable us to automate this option through the openshift installer byproviding it through
the install-config at installation time.