Network config in examples for openSUSE should be 1, not 2
The syntax follows version 1 https://cloudinit.readthedocs.io/en/latest/topics/network-config-format-v1.html

The issue is that, if you use version 2, you don't get the network
configured. However, this is not visible because we are running the
firstboot setup which will set the network config.

The issue is visible if you use an image without firstboot setup
enabled.

fixes #827

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.de>