Support TPM devices
"TPM (Trusted Platform Module) is a computer chip (microcontroller) that
can securely store artifacts used to authenticate the platform (your PC
or laptop). These artifacts can include passwords, certificates, or
encryption keys. A TPM can also be used to store platform measurements
that help ensure that the platform remains trustworthy. Authentication
(ensuring that the platform can prove that it is what it claims to be)
and attestation (a process helping to prove that a platform is
trustworthy and has not been breached) are necessary steps to ensure
safer computing in all environments."[1]

A TPM device is required by Windows 11 and it is also useful for testing
TPM support and can be used for protecting secrets in guests.

[1] https://trustedcomputinggroup.org/resource/trusted-platform-module-tpm-summary/
[2] https://docs.microsoft.com/en-us/windows/whats-new/windows-11-requirements

Fix #888