Remove NetworkUpdate workaround
The libvirt version check which is currently to decide whether to swap
libvirt.NetworkUpdate arguments is not enough. The fix was backported to
older libvirt in RHEL/CentOS (for example in
libvirt-6.0.0-37.1.module+el8.5.0+13858+39fdc467 ), and the current
version checking code would be swapping the arguments when it's no
longer needed.

This commit uses NetworkUpdateCompat which was recently introduced in
digitalocean/go-libvirt, and which takes care of this by checking
libvirt connection features.