Fix copy of the URL in RemoteName
Fixes #1040

The copy performed in RemoteName was shallow and this corrupted the
original URI that is used during the Dial() call made by the go-libvirt
library.

The issue always existed but did not manifest when libvirt.New(conn) was
used because the connection was established before the call to
RemoteName. The problem manifested when the switch to
libvirt.NewWithDialer was made, because the dialing is now made during
the call to ConnectToURI, which receives as paramater the result of the
call to RemoteName.