remove updated nodeinfo - leave existing implementation intact
this is related to the conversation in issue #1074. The original implementation
of nodeinfo (as released in PR #1042) has a bug - which may very well be easily
fixed - however the current PR and branch specifically addressed that issue by
calling virConnectGetCapabilities instead of virNodeGetInfo. Instead of doing
this, a new feature "host_capabilities" is added, and the nodeinfo code is left
unchanged (it may be addressed in a separate PR)