bugfix: be platform independent when dealing with filenames
previously this code would hand back the path as specified by the caller if it
failed (a path which can be a user specified string). However, this would mean
that the returned entity would be a string or a filepath depending on the
codepath. This behaviour previously existed in the codebase, however it is not
clear if it caused any issues anywhere in production. This PR, including the
associated unit test shows that under at least some circumstances (github ci)
the outcomes are unexpected.